classlogger-server/requirements.txt
flask==3.0.0
flask-cors==4.0.0
numpy==1.24.3
opencv-python==4.8.1.78
Pillow==10.1.0
firebase-admin==6.2.0
facenet-pytorch==2.5.3
torch==2.1.0
torchvision==0.16.0